refactor: 런닝 infra jpa 패키지 추가 #168
