+++ src/main/java/com/runky/running/infra/jpa/RunningJpaRepository.java
@@ -1,4 +1,4 @@
-package com.runky.running.infra;
+package com.runky.running.infra.jpa;
 
 import java.time.LocalDateTime;
 import java.util.Collection;
