+++ src/main/java/com/runky/running/infra/jpa/RunningRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.runky.running.infra;
+package com.runky.running.infra.jpa;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
