+++ src/main/java/com/runky/running/infra/jpa/RunningTrackJpaRepository.java
@@ -1,4 +1,4 @@
-package com.runky.running.infra;
+package com.runky.running.infra.jpa;
 
 import java.util.Optional;
